Badgetの入力フォームの作成

+++ db/migrate/20180125235808_create_expenses.rb
@@ -1,7 +1,7 @@
 class CreateExpenses < ActiveRecord::Migration[5.1]
   def change
     create_table :expenses do |t|
-      t.money :amount
+      t.integer :amount
       t.date :date
       t.string :note
       t.references :user, foreign_key: true
