+++ db/migrate/20180204083633_create_badgets.rb
@@ -1,7 +1,7 @@
 class CreateBadgets < ActiveRecord::Migration[5.1]
   def change
     create_table :badgets do |t|
-      t.money :amount
+      t.integer :amount
       t.references :user, foreign_key: true
       t.references :category, foreign_key: true
